refactor(domain): move out logic for removing a monitor to new `RemoveMonitorCommand`
