+++ GlazeWM.Domain/Monitors/Commands/RemoveMonitorCommand.cs
@@ -1,0 +1,14 @@
+﻿using GlazeWM.Infrastructure.Bussing;
+
+namespace GlazeWM.Domain.Monitors.Commands
+{
+  public class RemoveMonitorCommand : Command
+  {
+    public Monitor MonitorToRemove { get; set; }
+
+    public RemoveMonitorCommand(Monitor monitorToRemove)
+    {
+      MonitorToRemove = monitorToRemove;
+    }
+  }
+}
